fix: add manually-generated TLS cert
Automatic generation (and renewal) of TLS certificates (relying
on cert-manager and Let's Encrypt) does not work due to the
existence of an EnvoyFilter that blocks requests during
Let's Encrypt's challanges. Also a wildcard certificate must be
generated to properly deploy inference services within the platform.
This would require a DNS challenge and, consequently, some configuration
to be done in the DNS server managing platform's domain.

This complexity is left aside by generating a wildcard certificate for
all platform's endpoints. One drawback: certificate renewal is totally
manual from now on.

Documentation is updated accordingly.